Fix id type; better migration script

+++ db/warehouse/migrate/20231114235529_add_enrollment_to_locations.rb
@@ -1,7 +1,16 @@
 class AddEnrollmentToLocations < ActiveRecord::Migration[6.1]
   def change
     # Run the following manually after this is deployed
-    # ClientLocationHistory::Location.find_each { |clh| clh.update(enrollment_id: clh.source&.enrollment_id) }
-    add_column :clh_locations, :enrollment_id, :integer, index: true
+    # ClientLocationHistory::Location.find_each do |clh|
+    #   source = clh.source
+    #   id = if source.is_a?(GrdaWarehouse::Hud::Enrollment)
+    #     source.id
+    #   else
+    #     source.enrollment_id
+    #   end
+    #   clh.update(enrollment_id: id)
+    # end
+
+    add_column :clh_locations, :enrollment_id, :bigint, index: true
   end
 end
